pyproject.toml
[project]
name = "quack-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest>=7.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
